+++ ptypy/engines/projectional.py
