+++ templates/minimal_prep_and_run_DM_pycuda.py
@@ -6,7 +6,7 @@ of actual data. It uses the test Scan class
 
 from ptypy.core import Ptycho
 from ptypy import utils as u
-from ptypy.accelerate.cuda_pycuda.engines import DM_pycuda
+from ptypy.accelerate.cuda_pycuda.engines import projectional_pycuda
 p = u.Param()
 
 # for verbose output
