+++ templates/minimal_prep_and_run_DM_pycuda_stream.py
@@ -6,7 +6,7 @@ of actual data. It uses the test Scan class
 
 from ptypy.core import Ptycho
 from ptypy import utils as u
-from ptypy.accelerate.cuda_pycuda.engines import DM_pycuda_stream, DM_pycuda_streams, DM_pycuda
+from ptypy.accelerate.cuda_pycuda.engines import projectional_pycuda_stream
 p = u.Param()
 
 # for verbose output
